requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.4
psycopg-pool==3.2.4
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
pymongo==4.10.1
pytest==8.3.4
pytest-cov==4.1.0
mongomock==4.3.0
pydantic-settings==2.8.1
requests==2.32.3